# services/secret-service/doc/openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Secrets Service",
    "description" : "Secrets service is used in the Open Integration Hub for managing secrets used in integration flows.",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "https://skm.example.com"
  } ],
...
  "security": [
    {
      "BearerToken": []
    }
  ],
  "paths" : {
...
    "/api/v1/secrets/" : {
      "get" : {
        "tags" : [ "Secrets" ],
        "summary" : "Retrieve all secrets",
        "description" : "Retrieve all secrets belonging to the current user.",
        "operationId" : "getSecrets",
        "parameters": [
          {
            "$ref": "#/components/parameters/pageSize"
          },
          {
            "$ref": "#/components/parameters/pageNumber"
          }
        ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SecretEntry"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/500ServerError"
          }
        }
      },
      "post" : {
        "tags" : [ "Secrets" ],
        "summary" : "Create a secret",
        "description" : "Create a secret",
        "operationId" : "createSecret",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MutableSecret"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Secret"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/401Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/500ServerError"
          }
        }
      }
    },
    "/api/v1/secrets/{secretId}" : {
      "get" : {
        "tags" : [ "Secrets" ],
        "summary" : "Find secret by ID",
        "description" : "Returns a secret with given ID",
        "operationId" : "getSecretById",
        "parameters" : [ {
          "name" : "secretId",
          "in" : "path",
          "description" : "ID of secret to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          },
          "example": "6170a3c465dbff001297f235"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Secret"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/401Unauthorized"
          },
          "404" : {
            "description" : "Secret with given ID not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "Secrets" ],
        "summary" : "Delete a secret",
        "description" : "Delete a secret",
        "operationId" : "deleteSecret",
        "parameters" : [ {
          "name" : "secretId",
          "in" : "path",
          "description" : "ID of secret to delete",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          },
          "example": "6170a3c465dbff001297f235"
        } ],
        "responses" : {
          "204" : {
            "description" : "The secret was deleted successfully"
          },
          "404" : {
            "$ref": "#/components/responses/404NotFound"
          }
        }
      },
      "patch" : {
        "tags" : [ "Secrets" ],
        "summary" : "Update a secret",
        "description" : "Update a secret",
        "operationId" : "updateSecret",
        "parameters" : [ {
          "name" : "secretId",
          "in" : "path",
          "description" : "ID of secret to update",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MutableSecret"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Secret"
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "$ref": "#/components/responses/400BadRequest"
          }
        }
      }
    },
    "/api/v1/secrets/{id}/revoke": {
      "post": {
        "tags": ["Secrets"],
        "summary": "Revoke a client credentials token",
        "description": "Revokes the access token for a client credentials secret and deletes the secret. Only works for OA2_CLIENT_CREDENTIALS type secrets.",
        "operationId": "revokeClientCredentialsToken",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the secret to revoke"
          }
        ],
        "responses": {
          "204": {
            "description": "Token successfully revoked and secret deleted"
          },
          "400": {
            "description": "Invalid request - secret is not of type OA2_CLIENT_CREDENTIALS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401Unauthorized"
          },
          "404": {
            "description": "Secret or associated auth client not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during revocation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Owner" : {
        "required" : [ "id", "type" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "example": "617082258803c70031f85b26"
          },
          "type" : {
            "type" : "string",
            "description" : "Type of Owner entity, e.g. GROUP, TENANT, USER, etc.",
            "example": "USER"
          }
        }
      },
...
      "Meta" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "example": 1
          },
          "perPage" : {
            "type" : "integer",
            "example": 10
          },
          "total" : {
            "type" : "integer",
            "example": 15
          },
          "totalPages" : {
            "type" : "integer",
            "example": 2
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "MutableSecret" : {
        "required" : [ "name", "type", "owners" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "description" : "Human readable secret name",
            "example" : "My Secret"
          },
          "type" : {
            "type" : "string",
            "enum" : ["API_KEY", "OA1_TWO_LEGGED", "OA1_THREE_LEGGED", "OA2_AUTHORIZATION_CODE", "SIMPLE", "MIXED", "SESSION_AUTH"]
          },
          "value" : {
            "discriminator" : {
              "propertyName" : "type",
              "mapping": {
                "SIMPLE": "#/components/schemas/SimpleSecret",
                "MIXED": "#/components/schemas/MixedSecret",
                "API_KEY": "#/components/schemas/ApiKeySecret",
                "SESSION_AUTH": "#/components/schemas/SessionAuthSecret",
                "OA1_TWO_LEGGED": "#/components/schemas/OA1TwoLeggedSecret",
                "OA1_THREE_LEGGED": "#/components/schemas/OA1ThreeLeggedSecret",
                "OA2_AUTHORIZATION_CODE": "#/components/schemas/OA2AuthorizationCodeSecret",
                "OA2_PASSWORD": "#/components/schemas/OA2PasswordSecret",
                "OA2_CLIENT_CREDENTIALS": "#/components/schemas/OA2ClientCredentialsSecret"
              }
            },
            "oneOf" : [ {
              "$ref": "#/components/schemas/SimpleSecret"
            }, {
              "$ref": "#/components/schemas/MixedSecret"
            }, {
              "$ref": "#/components/schemas/ApiKeySecret"
            }, {
              "$ref": "#/components/schemas/SessionAuthSecret"
            }, {
              "$ref": "#/components/schemas/OA1TwoLeggedSecret"
            }, {
              "$ref": "#/components/schemas/OA1ThreeLeggedSecret"
            }, {
              "$ref": "#/components/schemas/OA2AuthorizationCodeSecret"
            }, {
              "$ref": "#/components/schemas/OA2PasswordSecret"
            }, {
              "$ref": "#/components/schemas/OA2ClientCredentialsSecret"
            } ]
          },
          "owners" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Owner"
            }
          },
          "tenant": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "lockedAt" : {
            "type" : "string",
            "description" : "Datetime (UTC) when the secret was locked. Relevant for 3legged-oauth-2 during acesstoken updates.",
            "format" : "date-time"
          },
          "encryptedFields": {
            "type": "array",
            "items": {
            }
          },
          "mixedProperties": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "Secret" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/MutableSecret"
        }, {
          "required" : [ "createdAt", "_id" ],
          "type" : "object",
          "properties" : {
            "_id" : {
              "type" : "string"
            },
            "createdAt" : {
              "type" : "string",
              "description" : "Secret creation time",
              "format" : "date-time"
            },
            "updatedAt" : {
              "type" : "string",
              "description" : "Secret update time",
              "format" : "date-time"
            },
            "value": {
              "type": "object",
              "additionalProperties": true
            }
          }
        } ]
      },
      "SimpleSecret" : {
        "required" : [ "passphrase", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "passphrase" : {
            "type" : "string"
          }
        }
      },
      "MixedSecret" : {
        "required": ["payload"],
        "type" : "object",
        "properties": {
          "payload": {
            "type": "string"
          }
        }
      },
      "ApiKeySecret": {
        "required": ["key"],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "headerName": {
            "type": "string"
          }
        }
      },
      "SessionAuthSecret": {
        "required": ["authClientId"],
        "type": "object",
        "properties": {
          "authClientId": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          },
          "inputFields": {}
        }
      },
      "OA1TwoLeggedSecret": {
        "type": "object",
        "properties": {
          "consumerKey": {
            "type": "string",
            "required": true,
            "description": "The Consumer Key may also be known as 'Client Key', 'Consumer Identifier', or 'API Key'"
          },
          "consumerSecret": {
            "type": "string",
            "required": true,
            "description": "The Consumer Secret may also be known as 'Client Secret', 'Shared Secret', or 'API Secret'"
          },
          "signatureMethod": {
            "type": "string",
            "required": true,
            "description": "HMAC-SHA1, HMAC-SHA256, etc."
          },
          "version": {
            "type": "string",
            "default": "1.0"
          },
          "accessToken": {
            "type": "string"
          },
          "accessTokenSecret": {
            "type": "string",
            "description": "Optional depending on the requirements of the endpoint, may also be called 'Token Secret'"
          },
          "realm": {
            "type": "string",
            "description": "Optional depending on the requirements of the endpoint"
          },
          "expiresAt": {
            "type": "string",
            "description": "OAuth1 access tokens do not typically expire, but this field may be useful for some implementations"
          }
        }
      },
      "OA1ThreeLeggedSecret" : {
        "required" : [ "authClientId", "accessToken", "accessTokenSecret" ],
        "type" : "object",
        "properties" : {
          "authClientId" : {
            "type" : "string",
            "description" : "Id of the auth client this secret was created with"
          },
          "accessToken" : {
            "type" : "string"
          },
          "accessTokenSecret" : {
            "type" : "string"
          },
          "scope": {
            "type": "string"
          },
          "expires": {
            "type": "string"
          },
          "externalId": {
            "type": "string"
          }
        }
      },
      "OA2AuthorizationCodeSecret" : {
        "required" : ["authClientId", "accessToken", "expires"],
        "type" : "object",
        "properties" : {
          "authClientId" : {
            "type" : "string",
            "description" : "Id of the auth client this secret was created with"
          },
          "refreshToken" : {
            "type" : "string"
          },
          "accessToken" : {
            "type" : "string"
          },
          "scope": {
            "type": "string"
          },
          "expires" : {
            "type" : "string",
            "description" : "Date object in UTC",
            "format" : "date-time"
          },
          "externalId": {
            "type": "string"
          }
        }
      },
      "OA2PasswordSecret" : {
        "required" : ["authClientId", "accessToken", "tokenType"],
        "type" : "object",
        "properties" : {
          "authClientId" : {
            "type" : "string",
            "description" : "Id of the auth client this secret was created with"
          },
          "refreshToken" : {
            "type" : "string"
          },
          "accessToken" : {
            "type" : "string"
          },
          "tokenType": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "expires" : {
            "type" : "string",
            "description" : "Date object in UTC",
            "format" : "date-time"
          },
          "fullResponse": {
            "type": "string",
            "description": "The full success response from the token endpoint (stored encrypted)"
          }
        }
      },
      "SecretEntry": {
        "type": "object",
        "required": ["_id", "name", "type", "owner"],
        "properties": {
          "_id": {
            "type": "string",
            "example": "61708255c0587c002c49f68a"
          },
          "name": {
            "type": "string",
            "example": "My Simple Secret"
          },
          "type": {
            "type": "string",
            "example": "SIMPLE"
          },
          "owners": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Owner"
            }
          }
        }
      }
    },
    "parameters": {
      "pageSize": {
        "name": "page[size]",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 10
        },
        "description": "The number of results to return per page",
        "required": false,
        "style": "deepObject"
      },
      "pageNumber": {
        "name": "page[number]",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1
        },
        "description": "The page number to return (not zero indexed)",
        "required": false
      }
    },
    "responses": {
      "400BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "401Unauthorized": {
        "description": "Unauthorized operation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "404NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "500ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "securitySchemes" : {
      "BearerToken": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}